Update message e2e tests to use seeder function instead of cli
The e2e tests file was using a function that executed a cli command. While that works, why not simply use the direct function that imports the seeds, rather than using a cli command to do exactly that?

Besides, calling the cli command wouldn't work for the test environment, as the function didn't specify which environment to send to the cli. This prevented the fixtures to be loaded correctly, failing the tests.