Remove .npmrc to promote global loglevel usage instead
loglevel=silent is interesting as in I (V-ed) do not want to see npm errors, as I find them obnoxious. However, they are useful in many situations, especially in CI's outputs, or for those that prefer to have them.

If you want this behavior back, simply add a .npmrc file in your HOME directory with the same content as removed in this commit, aka `loglevel=silent`.

https://docs.npmjs.com/cli/v7/configuring-npm/npmrc