Buff environment variables by validating its content
This uses `nest-typed-config` for a much better experience than Nest's own @nest/config package. Types are important yo!

The only downside with this is that you need to comment out empty keys, as they are parsed as `''`, which I couldn't map to be a default value because `class-validator` hates me or something