Add launch setup to run without watching
This allows to write tests much more easily since the server will not restart when spec files changes, but will still be debuggable in vscode.

I do not exactly _like_ this solution, I wished `tsc --watch` would respect the exclude option, but we have to deal with what we have all in all.