Tweak test environment to not run generators on push
This fixes the worst offender : whenever a test ran, the generators would run and trigger a server reload if watching.

This was especially noticeable when running an extension that monitors your tests, as it would restart the server randomly because of test monitoring.

This gives the best of both worlds : no watch confiiguration, but the ability to run the watch coonfig while having tests run in the background!