Fix fixture-loader not exiting if an error occured
If a seeder error occured, the whole function stopped at that point, making it impossible to disconnect the prisma database from within this function.

A try catch as been setup (with the missing await keyword) to remedy this.

Also, the seed results are sent to the resulting promise. The resulting promise will throw if an error occurred with the content of the latter.